+++ lib/include/openamp/hil.h
@@ -392,7 +392,7 @@ void hil_shutdown_cpu(struct hil_proc *proc);
  * returns -  status of function execution
  *
  */
-int hil_get_firmware(char *fw_name, unsigned int *start_addr,
+int hil_get_firmware(char *fw_name, uintptr_t *start_addr,
 		     unsigned int *size);
 
 /**
